# src/identity/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Role": {
				"enum": [
					0,
					1
				],
				"type": "number"
			},
			"UserStatus": {
				"enum": [
					"active",
					"blocked"
				],
				"type": "string"
			},
			"UserDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/UserStatus"
					}
				},
				"required": [
					"id",
					"email",
					"role",
					"createdAt",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PagedResult_UserDto-Array_": {
				"properties": {
					"result": {
						"items": {
							"$ref": "#/components/schemas/UserDto"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"result",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "identity",
		"version": "1.0.0",
		"description": "identity service",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Aleksei Popriadukhin"
		}
	},
	"paths": {
...
		"/api/v1/user/block": {
			"put": {
				"operationId": "BlockUser",
				"responses": {
					"204": {
						"description": "NO_CONTENT"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/v1/user/get": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PagedResult_UserDto-Array_"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"default": "ASC",
							"type": "string",
							"enum": [
								"ASC",
								"DESC"
							]
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"default": "id",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "searchTerm",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}